Split behaviors into individual module files. Fixes #3331.
