+++ kivy/factory_registers.py
@@ -102,12 +102,12 @@ r('PageLayout', module='kivy.uix.pagelayout')
 r('Accordion', module='kivy.uix.accordion')
 r('AccordionItem', module='kivy.uix.accordion')
 r('Button', module='kivy.uix.button')
-r('ButtonBehavior', module='kivy.uix.behaviors')
-r('ToggleButtonBehavior', module='kivy.uix.behaviors')
-r('DragBehavior', module='kivy.uix.behaviors')
-r('FocusBehavior', module='kivy.uix.behaviors')
-r('CompoundSelectionBehavior', module='kivy.uix.behaviors')
-r('CodeNavigationBehavior', module='kivy.uix.behaviors')
+r('ButtonBehavior', module='kivy.uix.behaviors.button')
+r('ToggleButtonBehavior', module='kivy.uix.behaviors.togglebutton')
+r('DragBehavior', module='kivy.uix.behaviors.drag')
+r('FocusBehavior', module='kivy.uix.behaviors.focus')
+r('CompoundSelectionBehavior', module='kivy.uix.behaviors.compoundselection')
+r('CodeNavigationBehavior', module='kivy.uix.behaviors.codenavigation')
 r('Bubble', module='kivy.uix.bubble')
 r('BubbleButton', module='kivy.uix.bubble')
 r('Camera', module='kivy.uix.camera')
